SR-13: moved the image excoprt file
moved the image export file to the theme/styles folder
