--- src/helpers/icons.ts
@@ -1,140 +1,0 @@
-import IconClock from '../assets/images/svg/clock.svg'
-import IconAlert from '../assets/images/svg/alert.svg'
-import IconArrowDownward from '../assets/images/svg/arrow-downward.svg'
-import IconArrowLeft from '../assets/images/svg/arrow-left.svg'
-import IconArrowRight from '../assets/images/svg/arrow-right.svg'
-import IconArrowUpward from '../assets/images/svg/arrow-upward.svg'
-import IconAtEmail from '../assets/images/svg/at-email.svg'
-import IconAward from '../assets/images/svg/award.svg'
-import IconBarEmptyLine from '../assets/images/svg/bar-empty-line.svg'
-import IconBarOneLine from '../assets/images/svg/bar-one-line.svg'
-import IconBarTwoLine from '../assets/images/svg/bar-two-line.svg'
-import IconBellTwo from '../assets/images/svg/bell-2.svg'
-import IconBell from '../assets/images/svg/bell.svg'
-import IconBook from '../assets/images/svg/book.svg'
-import IconCheckRound from '../assets/images/svg/check-round.svg'
-import IconCheckSmall from '../assets/images/svg/check-small.svg'
-import IconCheckMark from '../assets/images/svg/checkmark.svg'
-import IconClockOne from '../assets/images/svg/clock1.svg'
-import IconClose from '../assets/images/svg/close.svg'
-import IconCoin from '../assets/images/svg/coin.svg'
-import IconCrossSmall from '../assets/images/svg/cross-small.svg'
-import IconDefaultCheckbox from '../assets/images/svg/default-checkbox.svg'
-import IconEdit from '../assets/images/svg/edit.svg'
-import IconEyeClosed from '../assets/images/svg/eye-closed.svg'
-import IconEyeOpen from '../assets/images/svg/eye-open.svg'
-import IconFiltersTwo from '../assets/images/svg/filters-2.svg'
-import IconFilters from '../assets/images/svg/filters.svg'
-import IconFlagOne from '../assets/images/svg/flag-1.svg'
-import IconFlash from '../assets/images/svg/flash.svg'
-import IconFunnel from '../assets/images/svg/funnel.svg'
-import IconGrid from '../assets/images/svg/grid.svg'
-import IconHeart from '../assets/images/svg/heart.svg'
-import IconHome from '../assets/images/svg/home.svg'
-import IconLink from '../assets/images/svg/link.svg'
-import IconList from '../assets/images/svg/list.svg'
-import IconLogout from '../assets/images/svg/logout.svg'
-import IconMail from '../assets/images/svg/mail.svg'
-import IconMenuArrow from '../assets/images/svg/menu-arrow.svg'
-import IconMenu from '../assets/images/svg/menu.svg'
-import IconMessageCircle from '../assets/images/svg/message-circle.svg'
-import IconMessage from '../assets/images/svg/message.svg'
-import IconMoon from '../assets/images/svg/moon.svg'
-import IconPin from '../assets/images/svg/pin.svg'
-import IconPlus from '../assets/images/svg/plus.svg'
-import IconQuestionCircle from '../assets/images/svg/question-circle.svg'
-import IconQuestion from '../assets/images/svg/question.svg'
-import IconResizeSmall from '../assets/images/svg/resize-small.svg'
-import IconSearch from '../assets/images/svg/search.svg'
-import IconSettings from '../assets/images/svg/settings.svg'
-import IconStar from '../assets/images/svg/star.svg'
-import IconSun from '../assets/images/svg/sun.svg'
-import IconTrophy from '../assets/images/svg/trophy.svg'
-import IconUkraine from '../assets/images/svg/ukraine.svg'
-import IconUK from '../assets/images/svg/united-kingdom.svg'
-import IconUser from '../assets/images/svg/user.svg'
-import IconUsers from '../assets/images/svg/users.svg'
-import IconWarningCircle from '../assets/images/svg/warning-circle.svg'
-import IconWarningRound from '../assets/images/svg/warning-round.svg'
-
-import ImgDone from '../assets/images/done-registration.png'
-import ImgHtml from '../assets/images/html.png'
-import ImgJavaScript from '../assets/images/javascript.png'
-import ImgJs from '../assets/images/js.png'
-import ImgNodeJs from '../assets/images/nodejs.png'
-import ImgNotFound from '../assets/images/not-found.png'
-import ImgReact from '../assets/images/react.png'
-import ImgCompany from '../assets/images/registration-company.png'
-import ImgUser from '../assets/images/registration-monitor-user.png'
-
-export const icons = {
-  clock: IconClock,
-  alert: IconAlert,
-  arrowDownward: IconArrowDownward,
-  arrowLeft: IconArrowLeft,
-  arrowRight: IconArrowRight,
-  arrowUpward: IconArrowUpward,
-  atEmail: IconAtEmail,
-  award: IconAward,
-  barEmptyLine: IconBarEmptyLine,
-  barOneLine: IconBarOneLine,
-  barTwoLine: IconBarTwoLine,
-  bellTwo: IconBellTwo,
-  bell: IconBell,
-  book: IconBook,
-  checkRound: IconCheckRound,
-  checkSmall: IconCheckSmall,
-  checkMark: IconCheckMark,
-  clockOne: IconClockOne,
-  close: IconClose,
-  coin: IconCoin,
-  crossSmall: IconCrossSmall,
-  defaultCheckbox: IconDefaultCheckbox,
-  edit: IconEdit,
-  eyeClosed: IconEyeClosed,
-  eyeOpen: IconEyeOpen,
-  filtersTwo: IconFiltersTwo,
-  filters: IconFilters,
-  flagOne: IconFlagOne,
-  flash: IconFlash,
-  funnel: IconFunnel,
-  grid: IconGrid,
-  heart: IconHeart,
-  home: IconHome,
-  link: IconLink,
-  list: IconList,
-  logout: IconLogout,
-  mail: IconMail,
-  menuArrow: IconMenuArrow,
-  menu: IconMenu,
-  messageCircle: IconMessageCircle,
-  message: IconMessage,
-  moon: IconMoon,
-  pin: IconPin,
-  plus: IconPlus,
-  questionCircle: IconQuestionCircle,
-  question: IconQuestion,
-  resizeSmall: IconResizeSmall,
-  search: IconSearch,
-  settings: IconSettings,
-  star: IconStar,
-  sun: IconSun,
-  trophy: IconTrophy,
-  cUkraine: IconUkraine,
-  cUK: IconUK,
-  user: IconUser,
-  users: IconUsers,
-  warningCircle: IconWarningCircle,
-  warningRound: IconWarningRound,
-}
-export const imgs = {
-  done: ImgDone,
-  html: ImgHtml,
-  javaScript: ImgJavaScript,
-  js: ImgJs,
-  nodeJs: ImgNodeJs,
-  notFound: ImgNotFound,
-  react: ImgReact,
-  company: ImgCompany,
-  user: ImgUser,
-}
